move map to home screen
