+++ University-advisor-web/Controllers/HomeController.cs
@@ -1,38 +1,38 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
+using University_advisor_web.Constants;
+using University_advisor_web.Interfaces;
 using University_advisor_web.Models;
-using Microsoft.AspNetCore.Http;
+
+// For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace University_advisor_web.Controllers
 {
     public class HomeController : Controller
     {
-        private readonly ILogger<HomeController> _logger;
+        private readonly ILogger _logger;
 
-        public HomeController(ILogger<HomeController> logger)
+        public HomeController(ILogger logger)
         {
             _logger = logger;
         }
 
+        [HttpGet]
         public IActionResult Index()
-        {
-            return View();
+        {   
+            return View(new HomeModel());
         }
-
-        public IActionResult Privacy()
+        [HttpPost, ValidateAntiForgeryToken]
+        public IActionResult Index(string address, double range)
         {
-            return View();
-        }
-
-        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public IActionResult Error()
-        {
-            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+            var model = new HomeModel();
+            model.Map = new MapModel(address, "Universities", range);
+            // Code to test if logging works correctly.
+            _logger.Log(Messages.nearbyUniversitiesDisplayed);
+            return View(model);
         }
     }
 }
