+++ University-advisor-web/Controllers/LogInController.cs
@@ -44,7 +44,7 @@ namespace University_advisor_web.Controllers
                 HttpContext.Session.SetInt32("UserId", model.UserId);
                 HttpContext.Session.SetInt32("UserUniversityId", model.UniversityId);
                 HttpContext.Session.SetInt32("UserCourseId", model.CourseId);
-                return View("../Home/Index", model);
+                return View("../Home/Index", new HomeModel(model));
             }
             else
             {
