--- University-advisor-web/Controllers/MapController.cs
@@ -1,38 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using University_advisor_web.Constants;
-using University_advisor_web.Interfaces;
-using University_advisor_web.Models;
-
-// For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-
-namespace University_advisor_web.Controllers
-{
-    public class MapController : Controller
-    {
-        private readonly ILogger _logger;
-
-        public MapController(ILogger logger)
-        {
-            _logger = logger;
-        }
-
-        [HttpGet]
-        public IActionResult Index()
-        {   
-            var model = new MapModel("Vilnius", "Universities");
-            return View(model);
-        }
-        [HttpPost, ValidateAntiForgeryToken]
-        public IActionResult Index(string address, double range)
-        {
-            var model = new MapModel(address, "Universities", range);
-            // Code to test if logging works correctly.
-            _logger.Log(Messages.nearbyUniversitiesDisplayed);
-            return View(model);
-        }
-    }
-}
