+++ University-advisor-web/Views/Home/Index.cshtml
@@ -1,33 +1,54 @@
-﻿@{
-    ViewData["Title"] = "Home Page";
+﻿@model HomeModel
+@{
+    ViewData["Title"] = "Home";
 }
-<link rel="stylesheet" href="~/css/HomeScreen.css" />
-<div class="home-container">
-    <img src="~/img/natural.jpg" style="width:100%" />
-    <div class="image-text">
-        <h1>University advisor</h1>
-        <h3>We are here to help you with your journey</h3>
-    </div>
-</div>
-<div class="features text-center">
-    <h1>Features</h1>
-    <div class="d-flex justify-content-center features-wrapper">
-        <div class="features-item">Reviews</div>
-        <div class="features-item">UniCompare</div>
-        <div class="features-item">Sign up</div>
-    </div>
-    <div class="d-flex justify-content-center features-wrapper">
-        <div class="features-item">About us</div>
-        <div class="features-item">Contact us</div>
-    </div>
-    <div class="d-flex justify-content-center features-wrapper">
-        <div class="features-item">Top university list</div>
-    </div>
-</div>
-<div class="home-container">
-    <img src="~/img/vu.png" style="width:100%" />
-    <div class="image-text">
-        <h1>Vilnius University c# course is amazing!</h1>
-        <h3>- some software engineering student</h3>
-    </div>
-</div>
+<html>
+<head>
+    <link rel="stylesheet" href="~/css/HomeScreen.css" />
+    <link rel="stylesheet" href="~/lib/leaflet/leaflet.css" />
+    <!-- Make sure you put this AFTER Leaflet's CSS -->
+    <script src="~/lib/leaflet/leaflet.js"></script>
+    <style>
+        #map {
+            height: 60vh;
+            z-index: 50;
+        }
+
+        #sendAddressForm {
+            margin-bottom: 20px;
+        }
+    </style>
+</head>
+<body>
+    <h1 class="home-title"><span>Find Universities</span></h1>
+
+    <form asd-action="SendAddress" asp-controller="Home" role="form" method="POST" enctype="multipart/form-data" id="sendAddressForm">
+        <div class="row align-items-end">
+            <div class="col-lg-4">
+                <label asp-for="Map.Address"></label>
+                <input asp-for="Map.Address" type="text" class="form-control" pattern="[a-zA-Z0-9]+" />
+            </div>
+            <div class="col-lg-4">
+                <label asp-for="Map.Range"></label>
+                <input asp-for="Map.Range" type="number" class="form-control" required="" pattern="[0-9]{1,5}" />
+            </div>
+            <div class="col-lg-4">
+                <input class="btn btn-lg btn-primary" type="submit" value="Find">
+            </div>
+        </div>
+    </form>
+    @{
+        var locationData = @Html.Raw(Newtonsoft.Json.JsonConvert.SerializeObject(Model.Map));
+    }
+    <input type="hidden" id="locationsData" value="" />
+    <script>
+        document.getElementById('locationsData').value = JSON.stringify(@locationData);
+    </script>
+
+    <div id="map"></div>
+
+    <script src="~/js/AddMarkersToMap.js" type="text/javascript"></script>
+
+
+</body>
+</html>
