--- University-advisor-web/Views/Map/Index.cshtml
@@ -1,52 +1,0 @@
-﻿@model MapModel
-@{
-    ViewData["Title"] = "Map";
-}
-<html>
-<head>
-    <link rel="stylesheet" href="~/lib/leaflet/leaflet.css" />
-    <!-- Make sure you put this AFTER Leaflet's CSS -->
-    <script src="~/lib/leaflet/leaflet.js"></script>
-    <style>
-        #map {
-            height: 60vh;
-            z-index: 50;
-        }
-        #sendAddressForm {
-            margin-bottom: 20px;
-        }
-    </style>
-</head>
-<body>
-    <h1>Find Universities</h1>
-
-    <form asd-action="SendAddress" asp-controller="Map" role="form" method="POST" enctype="multipart/form-data" id="sendAddressForm">
-        <div class="row align-items-end">
-            <div class="col-lg-4">
-                <label asp-for="Address"></label>
-                <input asp-for="Address" type="text" class="form-control" pattern="[a-zA-Z0-9]+" />
-            </div>
-            <div class="col-lg-4">
-                <label asp-for="Range"></label>
-                <input asp-for="Range" type="number" class="form-control" required="" pattern="[0-9]{1,5}" />
-            </div>
-            <div class="col-lg-4">
-                <input class="btn btn-lg btn-primary" type="submit" value="Find">
-            </div>
-        </div>
-    </form>
-    @{
-        var locationData = @Html.Raw(Newtonsoft.Json.JsonConvert.SerializeObject(Model));
-    }
-    <input type="hidden" id="locationsData" value="" />
-    <script>
-        document.getElementById('locationsData').value = JSON.stringify(@locationData);
-    </script>
-
-    <div id="map"></div>
-
-    <script src="~/js/AddMarkersToMap.js" type="text/javascript"></script>
-
-
-</body>
-</html>
