+++ University-advisor-web/Views/Shared/_Layout.cshtml
@@ -61,9 +61,6 @@
                         <li class="nav-item ml-3">
                             <a class="nav-link text-white effect-box" asp-area="" asp-controller="Review" asp-action="Index">Reviews</a>
                         </li>
-                        <li class="nav-item ml-3">
-                            <a class="nav-link text-white effect-box" asp-area="" asp-controller="Map" asp-action="Index">Map</a>
-                        </li>
                         <li class="nav-item">
                             <a class="nav-link text-white effect-box" asp-area="" asp-controller="Forum" asp-action="Index">Forum</a>
                         </li>
